src/main/java/ru/job4j/Counter.java
package ru.job4j;

public class Counter {
    public static int sumByEven(int start, int finish) {
        int sum = 0;
        for (int i = start; i <= finish; i++) {
            if (i % 2 == 0) {
                sum += i;
            }
        }
        return sum;
    }

    public static void main(String[] args) {
        System.out.println(sumByEven(0, 10));
        System.out.println(sumByEven(3, 8));
        System.out.println(sumByEven(-5, 5));
        System.out.println(sumByEven(5, 5));
    }
}

